Refactor: Flatten sparse directories and consolidate domain modules
- Flattened `lib/list`, `lib/text`, `lib/type`, and `lib/extractor`
- Combined single-function files into domain-specific modules
- Added `lib/errorReporter.js` to centralize error reporting
- Updated tests and references
- Ensured no regressions in `npm test`
