+++ lib/extractor.js
@@ -1,4 +1,4 @@
-const mustString = require('../../../type/mustString')
+const mustString = require('./type')
 const urlRegexSafe = require('url-regex-safe')
 
 /**
@@ -18,4 +18,21 @@ function extractURLsFromString(text) {
     return ret
 }
 
-module.exports = extractURLsFromString
+/**
+ * Extracts the video id from a youtube url
+ * @param {string} url
+ */
+function extractYoutubeIDFromURL(url) {
+    return mustString(url)
+    .replace(/https?:/, '')
+    .replace(/\//g, '')
+    .replace("youtu.be", '')
+    .replace("www.", '')
+    .replace("youtube.com", '')
+    .replace("watch?v=", '')
+}
+
+module.exports = {
+    extractURLsFromString,
+    extractYoutubeIDFromURL
+}
