--- lib/extractor/common/urlsFromString/README.md
@@ -1,10 +1,0 @@
-# Overview
-
-Get all the URLs in a string as a list of strings
-
-The regexp is from a external library that are more battle proven than a handwritten dubitous regexp
-
-# Usecases
-- HTML link parsing
-- Filtering links to save to something like pocket
-- Anything that needs you to get links from a string
