--- lib/extractor/youtube/idFromUrl/README.md
@@ -1,6 +1,0 @@
-# Overview
-
-Function to get the youtube video id from the video URL
-
-# Usecases
-- A telegram bot to like videos when you can't login your account
