+++ lib/list.js
@@ -4,7 +4,6 @@
  * @param {Number} blockSize How many items you want in the sublists
  * @returns {Array}
  */
-
 function segment(list, blockSize) {
     let ret = []
     while (list.length !== 0) {
@@ -15,4 +14,20 @@ function segment(list, blockSize) {
     return ret
 }
 
-module.exports = segment
+/**
+ * Get the difference between two lists
+ * @param {Array} a
+ * @param {Array} b
+ * @returns {[Array, Array]}
+ */
+function diffList(a, b) {
+    return [
+        a.filter((v) => !b.includes(v)),
+        b.filter((v) => !a.includes(v))
+    ]
+}
+
+module.exports = {
+    segment,
+    diffList
+}
