--- lib/list/README.md
@@ -1,7 +1,0 @@
-# Functions for lists/arrays
-
-Here are functions to work with lists/arrays
-
-# Usecases
-
-- Functions that get a array of something to do something that is not equivalent to map with a function
