--- lib/list/diff/README.md
@@ -1,9 +1,0 @@
-# Overview
-
-Get differences between two lists
-
-Returns [added, removed]
-
-# Usecases
-
-- Twitter followers difference
