--- lib/list/segment/README.md
@@ -1,8 +1,0 @@
-# Overview
-
-Function to group a `list` of values in clusters of `blockSize` values
-
-# Usecases
-
-- Internal usage
-- Telegram message segmentation (the limit is 4096 characters)
