+++ lib/text.js
@@ -1,5 +1,5 @@
-const segment = require('../../list/segment')
-const mustString = require('../../type/mustString')
+const { segment } = require('./list')
+const mustString = require('./type')
 
 /**
  * Splits by line a text to fit in n bytes blocks.
@@ -35,4 +35,31 @@ function splitSizeByLine(text, blockSize) {
     return blocks
 }
 
-module.exports = splitSizeByLine
+/**
+ *
+ * @param {string | Buffer} data
+ * @returns string
+ */
+function toBase64(data) {
+    let input = data
+    if (!Buffer.isBuffer(input)) {
+        input = Buffer.from(mustString(input))
+    }
+    return input.toString('base64')
+}
+
+/**
+ * Convert a base64 string to a buffer, use method toString to convert to a string
+ * @param {string} b64data
+ * @returns {Buffer}
+ */
+function fromBase64(b64data) {
+    const data = mustString(b64data)
+    return Buffer.from(data, 'base64')
+}
+
+module.exports = {
+    splitSizeByLine,
+    toBase64,
+    fromBase64
+}
