--- lib/text/README.md
@@ -1,7 +1,0 @@
-# Text/string
-
-There are functions to work with strings or text
-
-# Usecases
-- Text structure manipulation
-- Quick conversions between encodings like base64
