--- lib/text/base64/README.md
@@ -1,9 +1,0 @@
-# Overview
-Functions to encode/decode data from/to base64
-
-# Usecases
-- Decode pipedream body
-
-- Encode binary data through non binary compatible channels like messaging APIs
-
-- Get binary data from non binary compatible channels
