--- lib/text/base64/to/index.js
@@ -1,16 +1,0 @@
-const mustString = require('../../../type/mustString')
-
-/**
- * 
- * @param {string | Buffer} data 
- * @returns string
- */
-function toBase64(data) {
-    let input = data
-    if (!Buffer.isBuffer(input)) {
-        input = Buffer.from(mustString(input))
-    }
-    return input.toString('base64')
-}
-
-module.exports = toBase64
