--- lib/text/base64/to/index.test.js
@@ -1,16 +1,0 @@
-const toBase64 = require('.')
-
-const input = "teste"
-const output = "dGVzdGU="
-
-test("should encode string", () => {
-    expect(toBase64(input)).toBe(output)
-})
-
-test("should encode a Buffer", () => {
-    expect(toBase64(Buffer.from(input))).toBe(output)
-})
-
-test("should convert number to string", () => {
-    expect(toBase64(2)).toBe("Mg==")
-})
