--- lib/type/README.md
@@ -1,11 +1,0 @@
-# Type/validation
-
-There are functions for typing and validation of data structures
-
-# Usecases
-- Validators for types
-- Atomic validators for primitive types and structures
-- Type asserts
-
-# Note
-- There should be autoexplicative names
